General refactor

+++ src/core/io.h
@@ -4,8 +4,8 @@
 #include "memory.h"
 
 class IO : public Memory<256> {
-    [[nodiscard]] uint8_t read(size_t index) const override;
-    void write(size_t index, uint8_t value) override;
+    [[nodiscard]] uint8_t read(uint16_t index) const override;
+    void write(uint16_t index, uint8_t value) override;
 };
 
 #endif // IO_H
